+++ packages/expo-ui/android/src/main/java/expo/modules/ui/SyncSwitchView.kt
@@ -1,7 +1,6 @@
 package expo.modules.ui
 
 import androidx.compose.material3.Switch
-import androidx.compose.material3.SwitchDefaults
 import androidx.compose.runtime.Composable
 import expo.modules.kotlin.views.ComposeProps
 import expo.modules.kotlin.views.FunctionalComposableScope
@@ -9,7 +8,6 @@ import expo.modules.ui.state.ObservableState
 
 data class SyncSwitchProps(
   val isOn: ObservableState? = null,
-  val label: String? = null,
   val enabled: Boolean = true,
   val modifiers: ModifierList = emptyList()
 ) : ComposeProps
@@ -17,14 +15,18 @@ data class SyncSwitchProps(
 @Composable
 fun FunctionalComposableScope.SyncSwitchContent(props: SyncSwitchProps) {
   val state = props.isOn ?: return
-  val checked = (state.value as? Boolean) ?: false
 
   Switch(
-    checked = checked,
+    checked = state.binding(false),
     onCheckedChange = { newValue ->
       state.value = newValue
     },
-    modifier = ModifierRegistry.applyModifiers(props.modifiers, appContext, composableScope, globalEventDispatcher),
+    modifier = ModifierRegistry.applyModifiers(
+      props.modifiers,
+      appContext,
+      composableScope,
+      globalEventDispatcher
+    ),
     enabled = props.enabled
   )
 }